rename CategoryTags to CategoriesTags
