+++ src/styles/categories-tags.scss
@@ -1,4 +1,4 @@
-.category-tags {
+.categories-tags {
   font-size: 20px;
   margin-top: 12px;
