+++ src/styles/index.scss
@@ -5,5 +5,5 @@
 @import "post.scss";
 @import "post-card.scss";
 @import "pagination.scss";
-@import "category-tags.scss";
+@import "categories-tags.scss";
 @import "share-button.scss";
